Move krate::index to krate::search::search
